requirements.txt
numpy==1.24.3
opencv_python==4.8.0.76
openslide_python==1.3.0
paquo==0.6.1
Shapely==1.8.1
tqdm==4.64.1
